refactor: norminette 맞추기
