+++ kipark_jayoon_minishell/Makefile
@@ -59,12 +59,14 @@ SRCS_EXECUTION				:=	do_it_child.c \
 								do_it_parent.c \
 								execute_cmd.c \
 								init_curr_envp.c \
+								init_fd_by_redirection.c \
 								init_execve_args.c \
 								is_single_cmd.c \
 								safe_close.c \
 								safe_dup2.c \
 								safe_fork.c \
 								safe_pipe.c \
+								safe_open.c \
 								wait_all_child.c \
 								here_doc.c \
 								here_doc_add.c \
