+++ kipark_jayoon_minishell/srcs/execution/safe_open.c
@@ -1,22 +1,37 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   init_redir_iter.c                                  :+:      :+:    :+:   */
+/*   safe_open.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: jayoon <jayoon@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/10/02 14:23:32 by jayoon            #+#    #+#             */
-/*   Updated: 2022/10/06 11:21:22 by jayoon           ###   ########.fr       */
+/*   Created: 2022/10/06 10:55:36 by jayoon            #+#    #+#             */
+/*   Updated: 2022/10/06 11:27:25 by jayoon           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include <fcntl.h>
 #include "libft.h"
-#include "parser.h"
 
-void	init_redir_iter(t_parsing_list *l_parsing)
+int	safe_open(char *path, int oflag)
 {
-	l_parsing->redir_iter
-		= (t_redir_iter *)ft_safe_malloc(sizeof(t_redir_iter));
-	l_parsing->redir_iter->l_input = NULL;
-	l_parsing->redir_iter->l_output = NULL;
+	int	fd;
+
+	fd = 0;
+	if (oflag == O_RDONLY)
+	{
+		fd = open(path, oflag);
+		ft_check_error(E_SYSTEM_CALL, (ssize_t)fd);
+	}
+	else if (oflag == (O_TRUNC | O_WRONLY | O_CREAT))
+	{
+		fd = open(path, oflag, 0666);
+		ft_check_error(E_SYSTEM_CALL, (ssize_t)fd);
+	}
+	else
+	{
+		fd = open(path, oflag, 0666);
+		ft_check_error(E_SYSTEM_CALL, (ssize_t)fd);
+	}
+	return (fd);
 }
