+++ kipark_jayoon_minishell/srcs/syntax_analysis/free_parsing_list.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   free_parsing_list.c                                :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: kipark <kipark@student.42seoul.kr>         +#+  +:+       +#+        */
+/*   By: jayoon <jayoon@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/02 14:07:53 by jayoon            #+#    #+#             */
-/*   Updated: 2022/09/24 22:04:03 by kipark           ###   ########seoul.kr  */
+/*   Updated: 2022/10/06 11:17:31 by jayoon           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,6 +25,7 @@ static void	free_simple_cmd_list(t_simple_cmd *head)
 		temp = NULL;
 	}
 }
+
 static void	free_redir_chunk_list(t_redir_chunk *head)
 {
 	t_redir_chunk	*temp;
