+++ kipark_jayoon_minishell/srcs/syntax_analysis/is_redirection.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   is_redirection.c                                   :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jayoon <jayoon@student.42.fr>              +#+  +:+       +#+        */
+/*   By: jayoon <jayoon@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/19 20:19:27 by jayoon            #+#    #+#             */
-/*   Updated: 2022/09/19 20:26:57 by jayoon           ###   ########.fr       */
+/*   Updated: 2022/10/06 11:21:58 by jayoon           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	is_redirection(t_token_type type)
 {
-	if (type == T_INPUT_REDIR || type == T_HERE_DOC 
+	if (type == T_INPUT_REDIR || type == T_HERE_DOC
 		|| type == T_OUTPUT_REDIR || type == T_APPEND_REDIR)
 		return (1);
 	return (0);
